+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Ref: Updates to align with other SDKs (#1134)
+- Add setter for value on the `ExceptionDataBag` (#1100)
+- Add `Scope::removeTag` method (#1126)
 
 ## 3.0.4 (2020-11-6)